Add support for a shortened form of demolishing.
Previously to demolish the addressbook example two blocks would need to be used to access the person elements. This commit adds support for a shortened form though it requires the passed block to accept two arguments the first of which is the demolisher node with the appropriate scope.

The new shortened form looks like

    xml.addressbook.person do |person, idx|
      puts person.firstname
    end

from the original

    xml.addressbook do
      xml.person do
        puts person.firstname
      end
    end

The requirement for the shortened form to accept arguments is due to the way in which demolisher manages the stack of nodes through which is traversed.